requirements.txt
pytest~=8.4.1
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
pytest-base-url
allure-pytest
PyYAML~=6.0.2
pip~=24.3.1
requests~=2.32.4
allure-python-commons~=2.15.0
jsonpath
pymysql